% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_causalForest.R
\name{CF_analysis}
\alias{CF_analysis}
\title{Performs causal forest to find factors contributing to effect heterogeneity}
\usage{
CF_analysis(ite, htn_med_class_i, df)
}
\arguments{
\item{htn_med_class_i}{The anti-hypertensive medication class of interest}

\item{df}{The analytic dataset}
}
\value{
Generates the feature importance plots for each medication class and outcome, also generates split values used by the CF algorithm
}
\description{
Performs causal forest to find factors contributing to effect heterogeneity
}
\details{
This function loads a json file with all the study cohorts
that were created using ATLAS and then generates them in the
user specified OMOP CDM database
}
