% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_TMLEanalysis.R
\name{createPatientProfiles}
\alias{createPatientProfiles}
\title{Create patient profiles}
\usage{
createPatientProfiles(cont_var, cutpoints, df)
}
\arguments{
\item{cont_var}{continuous outcomes}

\item{cutpoints}{cutpoints for the continuous variables}

\item{df}{analytic dataset description}
}
\value{
The ITE predictions
}
\description{
Create patient profiles
}
\details{
Creates patient profiles in the analytic dataset
}
