% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{executeStudy}
\alias{executeStudy}
\title{Executes the study analysis given database details}
\usage{
executeStudy(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "htn_hte",
  outputpath,
  generateCohorts = T,
  extractingData = T,
  runStatisticsAnalysis = T,
  runTreatmentEffects = F
)
}
\arguments{
\item{connectionDetails}{connection details to the OMOP CDM database}

\item{cdmDatabaseSchema}{The schema with the OMOP CDM database}

\item{cohortDatabaseSchema}{A schema with read/write access that the cohort table will be written to}

\item{cohortTable}{The name of the cohort table}

\item{outputpath}{The location to save results to}

\item{generateCohorts}{Whether to run SQL that creates the cohorts}

\item{extractingData}{Whether to extract the study data using the cohorts}

\item{runStatisticsAnalysis}{Whether to run the statistical analysis}

\item{runTreatmentEffects}{Whether to run the treatment effects}
}
\value{
The cohorts are generated into the specified cohort schema and table
}
\description{
Executes the study analysis given database details
}
\details{
Specify the database connection and schema details and then the study will execute
}
