% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0a_codeToExtractAtlasCohorts.R
\name{extractCohortDefinitionSet}
\alias{extractCohortDefinitionSet}
\title{Download cohort defintitions from ATLAS}
\usage{
extractCohortDefinitionSet(saveLocation)
}
\arguments{
\item{connectionDetails}{The OMOP CDM database connection details}

\item{cdmDatabaseSchema}{The schema of the OMOP CDM data}

\item{cohortDatabaseSchema}{The schema of the database where the cohort table will be created}

\item{cohortTable}{the table name that will be created to contain the cohorts}

\item{incremental}{Only regenerate new cohorts (i.e., skip cohorts that were previously generated and are unchanged)}

\item{incrementalFolder}{Location to save a file that tracks what cohorts have been generated previously for incremental}
}
\value{
The cohorts are generated into the specified cohort schema and table
}
\description{
Download cohort defintitions from ATLAS
}
\details{
This function downloads cohort definitions from ATLAS
}
