% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d_cleanDataset.R
\name{generateAnalyticDataset}
\alias{generateAnalyticDataset}
\title{Clean and wrangle dataset for analysis}
\usage{
generateAnalyticDataset(data)
}
\arguments{
\item{data}{The data matrix generated by the generateCohort function}
}
\value{
The cohorts are generated into the specified cohort schema and table
}
\description{
Clean and wrangle dataset for analysis
}
\details{
This function loads a json file with all the study cohorts
that were created using ATLAS and then generates them in the
user specified OMOP CDM database
}
