% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohorts.R
\name{getData}
\alias{getData}
\title{Extract the data for the study}
\usage{
getData(
  connectionDetails,
  cdmDatabaseSchema,
  cdmDatabaseName,
  cohortDatabaseSchema,
  cohortTable = "htn_hte"
)
}
\arguments{
\item{connectionDetails}{The OMOP CDM database connection details}

\item{cdmDatabaseSchema}{The schema of the OMOP CDM data}

\item{cohortDatabaseSchema}{The schema of the database where the cohort table will be created}

\item{cohortTable}{the table name that will be created to contain the cohorts}

\item{cdmDatabaseName.}{A friendly name for theOMOP CDM database}
}
\value{
Extracts the data for the study and then converts into a sparse matrix
}
\description{
Extract the data for the study
}
\details{
This function extracts the data for the study
}
