% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTNHTE.R
\docType{package}
\name{HTNHTE-package}
\alias{HTNHTE}
\alias{HTNHTE-package}
\title{HTNHTE: Personalized Causal Inference Study}
\description{
A study to investigate personalised causal inference
}
\author{
\strong{Maintainer}: add name \email{addemail@here.com}

}
\keyword{internal}
