% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ITEestimation.R
\name{step1_ite_SL}
\alias{step1_ite_SL}
\title{Performs Q-estimation for counterfactual predictions}
\usage{
step1_ite_SL(med_class_i, df, outcome)
}
\arguments{
\item{med_class_i}{The anti-hypertensive medication class of interest}

\item{df}{The analytic dataset}

\item{outcome}{The outcome of interest}
}
\value{
The ITE predictions
}
\description{
Performs Q-estimation for counterfactual predictions
}
\details{
This function loads a json file with all the study cohorts
that were created using ATLAS and then generates them in the
user specified OMOP CDM database
}
