% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_TMLEanalysis.R
\name{TMLE_analysis_neg}
\alias{TMLE_analysis_neg}
\title{Performs TMLE for the negative control outcome in the overall population}
\usage{
TMLE_analysis_neg()
}
\arguments{
\item{patient_profile_list}{list of patient profiles created description}
}
\value{
The ITE predictions
}
\description{
Performs TMLE for the negative control outcome in the overall population
}
\details{
Performs TMLE for the negative control outcome in the overall population
}
