% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_TMLEanalysis.R
\name{TMLE_analysis}
\alias{TMLE_analysis}
\title{Performs TMLE over all patient profiles}
\usage{
TMLE_analysis(df, outcome, patient_profile_list, htn_med_list)
}
\arguments{
\item{outcome}{The outcome of interest}

\item{patient_profile_list}{list of patient profiles created description}

\item{htn_med_list}{The different treatments}
}
\value{
The ITE predictions
}
\description{
Performs TMLE over all patient profiles
}
\details{
Performs TMLE for all patient profiles for the main outcomes
}
